requirements.txt
brotli # automatically handle decompression
fastmcp>=3.0.0b1
opentelemetry-distro
opentelemetry-exporter-otlp

sqlalchemy

fastapi
uvicorn
aiocache
pydantic
async-lru

langchain-google-genai

pydantic-settings
python-dotenv

ruff